+++ src/Umbraco.Web.Common/Events/ActionExecutedEventArgs.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Umbraco.Web.BackOffice
+namespace Umbraco.Web.Common.Events
 {
     public class ActionExecutedEventArgs : EventArgs
     {
